split switch layers into its own file (to better match the python version)
