+++ Libraries/LLM/PhiMoE.swift
@@ -261,151 +261,3 @@ extension PhiMoEModel: LoRAModel {
         model.layers.map { ($0.selfAttn, ["q_proj", "v_proj"]) }
     }
 }
-
-// MARK: - SwitchGLU
-
-class SwitchGLU: Module {
-    @ModuleInfo(key: "gate_proj") var gateProj: SwitchLinear
-    @ModuleInfo(key: "up_proj") var upProj: SwitchLinear
-    @ModuleInfo(key: "down_proj") var downProj: SwitchLinear
-
-    let inputDims: Int
-    let hiddenDims: Int
-    let numExperts: Int
-    let activation: (MLXArray) -> MLXArray
-
-    init(
-        inputDims: Int,
-        hiddenDims: Int,
-        numExperts: Int,
-        activation: @escaping (MLXArray) -> MLXArray = MLXNN.silu,
-        bias: Bool = false
-    ) {
-        self.inputDims = inputDims
-        self.hiddenDims = hiddenDims
-        self.numExperts = numExperts
-        self.activation = activation
-
-        self._gateProj.wrappedValue = SwitchLinear(
-            inputDims: inputDims, outputDims: hiddenDims, numExperts: numExperts, bias: bias)
-        self._upProj.wrappedValue = SwitchLinear(
-            inputDims: inputDims, outputDims: hiddenDims, numExperts: numExperts, bias: bias)
-        self._downProj.wrappedValue = SwitchLinear(
-            inputDims: hiddenDims, outputDims: inputDims, numExperts: numExperts, bias: bias)
-
-        super.init()
-    }
-
-    func callAsFunction(_ x: MLXArray, _ indices: MLXArray) -> MLXArray {
-        let x = MLX.expandedDimensions(x, axes: [-2, -3])
-
-        let xUp = upProj(x, indices)
-        let xGate = gateProj(x, indices)
-        let xDown = downProj(activation(xGate) * xUp, indices)
-
-        return MLX.squeezed(xDown, axis: -2)
-    }
-}
-
-class SwitchLinear: Module, Quantizable {
-    @ModuleInfo(key: "weight") var weight: MLXArray
-    @ModuleInfo(key: "bias") var bias: MLXArray?
-
-    let inputDims: Int
-    let outputDims: Int
-    let numExperts: Int
-
-    init(inputDims: Int, outputDims: Int, numExperts: Int, bias: Bool = true) {
-        self.inputDims = inputDims
-        self.outputDims = outputDims
-        self.numExperts = numExperts
-
-        let scale = sqrt(1.0 / Float(inputDims))
-        self._weight.wrappedValue = MLXRandom.uniform(
-            low: -scale,
-            high: scale,
-            [numExperts, outputDims, inputDims]
-        )
-
-        if bias {
-            self._bias.wrappedValue = MLXArray.zeros([numExperts, outputDims])
-        }
-
-        super.init()
-    }
-
-    /// Initializer meant for subclasses to provide weight and bias arrays directly.
-    ///
-    /// This is used e.g. by ``QuantizedSwitchLinear`` to provide quantized weights and biases
-    /// rather than have ``SwitchLinear`` compute them.
-    init(
-        inputDims: Int, outputDims: Int, numExperts: Int,
-        weight: MLXArray, bias: MLXArray? = nil
-    ) {
-        self.inputDims = inputDims
-        self.outputDims = outputDims
-        self.numExperts = numExperts
-
-        self._weight.wrappedValue = weight
-        self._bias.wrappedValue = bias
-    }
-
-    func callAsFunction(_ x: MLXArray, _ indices: MLXArray) -> MLXArray {
-        let weightT = self.weight.swappedAxes(-1, -2)
-        var result = MLX.gatherMatmul(x, weightT, rhsIndices: indices)
-
-        if let bias = self.bias {
-            result = result + MLX.expandedDimensions(MLX.take(bias, indices), axis: -2)
-        }
-
-        return result
-    }
-
-    func toQuantized(groupSize: Int = 64, bits: Int = 4) -> Module {
-        QuantizedSwitchLinear(self, groupSize: groupSize, bits: bits)
-    }
-}
-
-class QuantizedSwitchLinear: SwitchLinear {
-    @ModuleInfo(key: "scales") var scales: MLXArray
-    @ModuleInfo(key: "biases") var biases: MLXArray
-
-    let groupSize: Int
-    let bits: Int
-
-    init(_ other: SwitchLinear, groupSize: Int = 64, bits: Int = 4) {
-        self.groupSize = groupSize
-        self.bits = bits
-
-        let (quantizedWeight, scales, biases) = MLX.quantized(
-            other.weight, groupSize: groupSize, bits: bits)
-
-        self._scales.wrappedValue = scales
-        self._biases.wrappedValue = biases
-
-        super.init(
-            inputDims: other.inputDims, outputDims: other.outputDims, numExperts: other.numExperts,
-            weight: quantizedWeight, bias: other.bias)
-
-        self.freeze()
-    }
-
-    override func callAsFunction(_ x: MLXArray, _ indices: MLXArray) -> MLXArray {
-        var result = MLX.gatherQuantizedMatmul(
-            x,
-            self.weight,
-            scales: self.scales,
-            biases: self.biases,
-            rhsIndices: indices,
-            transpose: true,
-            groupSize: self.groupSize,
-            bits: self.bits
-        )
-
-        if let bias = self.bias {
-            result = result + MLX.expandedDimensions(MLX.take(bias, indices), axis: -2)
-        }
-
-        return result
-    }
-}
